Add logger injection to CouchKVStore
This adds the necessary changes to allow a test (or something else)
to inject a logger object into a new CouchKVStore instance via the
KVStoreConfig. By default it currently uses the global logger.

When creating a KVStoreConfig the Logger can be set as follows:

    Logger logger;
    auto config = KVStoreConfig( .. ).setLogger(logger);

The purpose of this is to allow, in a future commit, verifying
logging output from injected errors into couchstore.

Change-Id: If0b91b52d89b4806f27634e7fdaace0cf5fb3a63
Reviewed-on: http://review.couchbase.org/62192
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>