MB-18974: Fix intermittant failure in test_expiration_on_warmup
This test attempts to check that after a restart, the expiry pager is
run and will remove items which have now expired. However, it checks
the 'ep_num_expiry_pager_runs' to determine when the expiry pager has
been run - except this stat is incremented at the /start/ of the
expiry pager Task, and hence there is a race condition whereby even if
this stat is non-zero it may not yet have visited all items.

Fix by instead checking the 'ep_expired_pager' stat which counts how
many items have been expired - and crucially is updated when the item
is actually expired.

Change-Id: I0bc3fb79b8bed78e88d1d6ed7fd468b113705846
Reviewed-on: http://review.couchbase.org/62234
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>