MB-18940: Make MemoryTracker::getInstance() thread-safe
The MemoryTracker::getInstance() function, used to obtain a pointer to
the Memory Tracker singleton is not thread-safe, and may result in two
(or more!)  MemoryTracker objects being created. This in turn can
result in multiple stats threads bring spawned, which at shutdown time
can cause a crash when trying to join threads which we no longer have
correct threadIds for.

Solve this problem by making getInstance() thread-safe, by making the
instance variable atomic and using double-locking when creating the
singleton.

(As an aside, in C++11 we /should/ be able to simplify this code
 significantly by using 'magic statics' - e.g.:

    MemoryTracker* MemoryTracker::getInstance() {
        static MemoryTracker instance;
        return &instance;
    }

 However they are /not/ supported in MSVC 2013 - see 'Magic statics'
 row in [Support For C++11 Features][1] - so we need to use a more
 involved approach for now).

[1]: https://msdn.microsoft.com/en-us/library/hh567368(v=vs.120).aspx#concurrencytable

Change-Id: Id52c3fcd5430a8726fac03a05ff602ea073b6084
Reviewed-on: http://review.couchbase.org/62286
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>