MB-18974: Fix intermittant failure in test_expiration_on_warmup (2)
There is an additional issue with the expiration_on_warmup test which
can cause it to fail intermittently - we increment the 'expired_pager'
statistic /before/ we actually expire the item. This means there is a
potential window in the test where the expired_pager statistic will
report '1', yet curr_items will not report 0 (as the item hasn't yet
been removed).

Fix by deferring the expired_pager statistic update until we have
actually expired the item.

Change-Id: Iab7bf30edbea049efccd0746d6208218e931c205
Reviewed-on: http://review.couchbase.org/62296
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>