MB-19033 Allow connections of the same name
It was thought that connections would not exist
in the connection map with the same name.

However the view code does create connections of
the same name.  As we index on the connection object
(cookie) as opposed to the name, we can relax the
constraints to allow connections of the same name.

Change-Id: I721c4d409d7f02119af534cbf1d887d9e65246c3
Reviewed-on: http://review.couchbase.org/62361
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>