MB-19033: Release cookie if newProducer/newConsumer fails
In the dcpOpen function we first call reserveCookie on a
connection object and then call newProducer/newConsumer
to create a new connection.

If the newProducer/newConsumer fails to create a new
connection and instead returns a nullptr we must call
releaseCookie before returning ENGINE_DISCONNECT.

A test is also added to ep_testsuite_dcp that provide a
regression test for the fix.

Change-Id: I1aceea01ae0e764f4118e4a5e5b29e2aa8ff30f0
Reviewed-on: http://review.couchbase.org/62515
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>