MB-18974: Fix intermittant failure in test_expiration_on_warmup (3)
There is an *another* issue with the expiration_on_warmup test which
can cause it to fail intermittently - investigation shows that the
curr_items stat is not synchronous with respect to expired_pager.

Specifically, for non-temporary, expired items we call
unlocked_softDelete (soft-marking the item as deleted in the
hashtable), and then call queueDirty to queue a deletion, and then
increment the expired stat. Only when that delete is actually
persisted and the deleted callback is invoked -
PeristenceCallback::callback(int&) - is curr_items finally
decremented.

Therefore we need to wait for the flusher to settle (i.e. delete
callback to be called) for the curr_items stat to be accurate.

This is hopefully the last patch needed to address this issue (however
I did say that last time :/)

Change-Id: Iaec44e4149c6fef549036870b31c9e0631f3949b
Reviewed-on: http://review.couchbase.org/62440
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>