MB-19086: Do not read from backfill if the scanBuffer is full
While backfilling, we want to limit the amount of bytes read in one
run of the backfill (for one vbucket). This commit address a bug in
that logic.

To test this we need to check how many times the backfill task runs.
To do this as part of the commit, code to read histogram stats in
ep_testsuite is added.

Change-Id: Ia5f653325583ebae32e1b858924c29327e035318
Reviewed-on: http://review.couchbase.org/62463
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>