MB-19184: Use AtomicUnorderedMap for PassiveStream::streams
Refactor the streamsMap in DcpConsumer to use the new
AtomicUnorderedMap class instead of a fixed array of pointers. While
the fixed array should be safe, it's unnecessarily inefficient in
terms of space as it always takes up 1024 * sizeof(RCPtr<>), which is
16KB on x86-64.

Moving to the new AtomicUnorderedMap gives us a single base
implementation and a memory footprint proportional to the number of
vbuckets in use.

Change-Id: Ie975b68725f7c3f012c26cfdedd44bf052d42e96
Reviewed-on: http://review.couchbase.org/62846
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>