MB-19205: Speed up audit tests
Previously each audit test would start and stop memcached and
this works pretty fast when running on your laptop, but under
valgrind this is not the case as it tries to track all memory
access for each startup etc.

This patch rewrites the logic so that each test case enable
and disable the audit framework and tries to read the
event from disk as soon as possible (with a 5 sec max
timeout). The audit daemon tries to store the stuff
to disk as fast as possible so we shouldn't have to wait
that long for it to hit the disk in the normal situation

Change-Id: Ic140806335e061c3f0f4a387711772b9e56d62a3
Reviewed-on: http://review.couchbase.org/63083
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>