Add synchronous EPStoreTests - GetKeyStats tests
Add tests for EPStore::getKeyStats, both for full eviction and value
eviction.

These tests use a new Test base class (EventuallyPersistentStoreTest)
and mocks which set up EPEngine / Store in a synchronous mode, so
normally background tasks can be explicitly triggered as desired.

This allows testing of things like a 2nd request for a non-resident
key (triggering EWOULDBLOCK) without the non-determinism of the
BGFetcher task running in the background and interfering.

Change-Id: Ib5cfe447fc9b0adb7c5e50c8778b890972f41078
Reviewed-on: http://review.couchbase.org/64023
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>