getKeyStats: Remove redundent 'bgfetch' parameter
EventuallyPersistentStore::getKeyStats takes a {bgfetch} parameter to
determine if a bgfetch should be performed as part of this
request. However this parameter is redundent - is is only applicable
if FULL is the current eviction policy, and all callers of the
function set it to true when FULL is selected.

Therefore remove the parameter to simply the code.

Change-Id: I09b0a211850326aca6ae9c91777c32eb8d0eddbe
Reviewed-on: http://review.couchbase.org/63186
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>