Don't hold the task lock during the completion message
The reason for this change is that we end up with
hacks in the callback that unlocks/relocks the mutex
before trying to access the connection (since that
would aqcuire the connections thread lock, and the
locking order would be the oposite when the thread is
created). This lock/unlock/lock/unlock bracketing just
made it harder (and more difficult) to understand.

Change-Id: I607ef2a133f854bdb8e8d1c3f42da58aeafae2fd
Reviewed-on: http://review.couchbase.org/63312
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>