Change AtomicUnorderedMap to pass by value
AtomicUnorderedMap was initially written to wrap all returned values
in a templated smart pointer type. This meant that it always had
ownership semantics. For the initial use-cases (DCP Consumer /
Producer maps) this was fine, however other use-cases may not want
ownership semantics, for example if a user holds a map of key -> raw
pointer type, and they can assume that the lifetime of the value type
outlives the map. Such a use-case could not previously use
AtomicUnorderedMap (as if you deleted the element it would potentially
destroy the object). Additionally, AtomicUnorderedMap couldn't be used
for simple value types where the overhead of smart pointer is
undesired (simple copying is fine).

Therefore modify the class to just hold the templated mapped_type
directly, deferring to the user if the value type happens to be a
pointer, smart or otherwise. All methods which previously returned a
PtrType now return a pair of {T, bool} where T is the value found (or
default-constructed if not found), and the bool denotes if T was
found. This is necessary as the null pointer type can no longer be
used to denote the absence of a value.

Change-Id: If82f10a3802d72eb3bf47f8d8b91d68c193cefdf
Reviewed-on: http://review.couchbase.org/63414
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>