MB-19688: Fix intermittent ep_perfsuite failure due to buggy string COW
ThreadSanitiser flagged that there is a data-race accessing the result
of get_str_stat(). It appears we are falling foul of an optimisation
that certain implementations[1] of std::string perform, that is to
provide a COW implementation. Thus we look like we are creating a copy
of an object, but under the covers a shared string is being
accessed. Note this is buggy behaviour under C++11, but GCC < 5 (using
it's COW string optimization) suffers from this issue.

The workaround is to explictly construct a new std::string which is a
copy of actual_stat_value.

Note: We've encountered the same issue elsewhere - see MB-18510 /
commit ea89a57047c147e3c7defa0b9a3bed44aafae02b.

[1]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=21334#c47

Change-Id: I14c14ce636becbc34d9963c8746a721cdc0cad1b
Reviewed-on: http://review.couchbase.org/63460
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>