Unify CouchKVStore::{getItemCount,getDbFileInfo}
CouchKVStore::getItemCount and getDbFileInfo do virtally the same
thing - open a couchstore file, read the header information and then
return some subset of that information to the caller.

Given this, we can implement both in terms of a new helper function -
getDbInfo - which removes all of the code duplication between the two
(mostly error-handling).

Change-Id: I8338e17f9e9cc31c7273949f9838bb23f618f3dc
Reviewed-on: http://review.couchbase.org/63527
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>