Remove default arguments from HashTable::unlocked_add
HashTable::unlocked_add() has four defaulted boolean arguments; which
can be confusing to the reader as to which option is set to
what.

Given that one of these arguments always have the same value
(storedVal is always true), make the flags explicit in the callers to
assist in refactoring / simplification.

Change-Id: I95defa89cfb2411624f39963780390ff2af56412
Reviewed-on: http://review.couchbase.org/63699
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>