Set NRU from Item instead of explicit parameter
When storing items into the HashTable there is an optional argument to
specify the nru (not recently used) value for a StoredValue. This is
inconsistent with all other properties, which are passed via the Item
object; additionally it complicates the various HashTable add/set
calls as they have a defaulted argument.

Change to use the Item's NRU when populating the HashTable.

Change-Id: I9a8ac8597b82c4ea897d450a35f581ad4e9a14ca
Reviewed-on: http://review.couchbase.org/63758
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>