MB-19501: ep_test_apis.cc: Wait for outstanding flushers to complete
There is an issue in how the testsuites wait for items to be flushed
to disk when in full eviction mode, which results in intermittent test
failures are the item count (curr_items) is under-reported.

The issue is that we only wait for the {{ep_queue_size}} stat to be
zero (assuming that means that all items are now on disk), however in
the case of full eviction we need to update the count of how many
items are in a vBucket (HashTable) based on how many are now on disk;
with is done after the disk commit - see
EventuallyPersistentStore::commit().

We also need to to wait for any outstanding flushes to disk to
complete - specifically so when in full eviction mode we have waited
for the item counts in each vBucket to be synced with the number of
items on disk.

Change-Id: Ibbffddd137fc17f1dfe80edcc35e177ee6915763
Reviewed-on: http://review.couchbase.org/63710
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>