MB-19504: Remove full-eviction variant of ep_perfsuite
The full eviction variant of ep_perfsuite is essentially a waste of
time (and prone to intermittent failures).

The perfsuite tests run with persistence disabled (see the call to
stop_persistence at the start of perf_latency. As a consequence the
tests sometimes fail, as items are essentially getting evicted from
memory (and put onto the disk queue) but the queue will never be
persisted, and hence item counts will not be correctly updated. (Note:
this relates to the issues found in MB-19501).

Additionally, even if the tests did work, given we disable persistence
there's essentially no value in the results output as they aren't
representative of what full eviction operations might cost (as no disk
overhead is included).  For these reasons I'm disabling the full
eviction mode of ep_perfsuite. While in the abstract it would be
useful to have numbers from this suite for full eviction, it requires
a reasonable amount of work to make work in a sensible way.

Change-Id: I450dc7297b25d3c6b09408688d5d7706bc26b5fe
Reviewed-on: http://review.couchbase.org/63712
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>