MB-19430: Handle temporary failure in BackfillDiskLoad::run()
memcached.log contains the following error message:

    "CouchKVStore::getDbFileInfo: Failed to open database file
    for vBucket = 0 rev = 1 with error:no such file:

The reason for this failure is a race condition between the thread
which creates couch-db and the thread which opens this file for read
operation (to get DbFileInfo).

To fix, if BackfillDiskLoad::run fails to obtain the item count for a
vbucket (due to the file not being ready yet), then snooze the task
and retry later.

Change-Id: I4613ef4716b0a1dccd5928d776e1f20ecdfe129e
Reviewed-on: http://review.couchbase.org/63929
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>