MB-19503: Unit test to demonstrate notifications can go missing.
The test serially interleaves ConnMap::notifyPausedConnection into the
path of ConnMap::notifyAllPausedConnections so that we show a second
notification goes missing because of the ordering of clearing
the Notifiable::notificationScheduled atomic bool.

Test is currently marked as disabled because the fix is coming
from downstream via merge commits.

Fix -> http://review.couchbase.org/#/c/63934/

Change-Id: Icd7348364c393f154ad5db9c741fc1616a121805
Reviewed-on: http://review.couchbase.org/63977
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: buildbot <build@couchbase.com>