MB-19607: Fix use-after-free in DestroyBucketThread::destroy
When printing the numeric ID of a connection, we convert it to a
std::string and then call c_str() on an rvalue. This results in us
having a pointer to a deleted object (as the std::string will be
deleted at the end of the assignment expression).

Solve by making the variable hold a concrete std::string object
instead of char*.

Change-Id: Ie58607f2907fbb9d2dd38c8815e6f917a4d597b6
Reviewed-on: http://review.couchbase.org/63976
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>