MB-19639: engine_testapp: Allow >1 attempt for a test to pass
We (unfortunately) have some tests driven by engine_testapp which
intermittently fail - for example ep_testsuite. While much has been
done to improve these test issues (particulary on newer branches), the
fact remains that some tests are flaky and need retrying to pass.

For commit-validation jobs, this means that the whole job will fail,
and has to be manually retried. Not only is the repeat job
time-consuming (CV job runtimes vary, but say ~10mins on average), but
it also requires human intervention to trigger the retry.

To attempt to improve this, this patch adds the option to attempt a
test multiple times before declaring it a failure. The number of
attempts can be controlled either via a new command-line argument:

    -a <attempts>    Maximum number of attempts for a test.

or via an environment variable:

    export TESTAPP_ATTEMPTS=5

The default attempts is 1 - which results in the same behaviour as
previously.

If a test passes on retry (OK AFTER RETRY), then it doesn't contribute
to the error count - and hence the overall testsuite status is
pass. Example output of running with attempts = 10:

    Running [0000/0005]: validate engine handle (couchstore)...(0 sec) FAIL
    (1 sec) FAIL
    (0 sec) OK AFTER RETRY
    Running [0001/0005]: test alloc limit (couchstore)...(0 sec) OK
    Running [0002/0005]: test total memory limit (couchstore)...(2 sec) OK
    Running [0003/0005]: test max_size changes (couchstore)...(0 sec) OK
    Running [0004/0005]: test whitespace dbname (couchstore)...(0 sec) OK
    # Passed 5 of 5 tests

    $ echo $?
    0

Change-Id: Id2480f9e92714009292a6695a6aee5877324e233
Reviewed-on: http://review.couchbase.org/64110
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>