MB-19557: Catch std::exception in McbpConnection::runEventLoop
In the event of some (unhandled) exception being thrown by code run on
the memcached frontend worker threads we catch it and print, closing
the current connection.

However, only std::invalid_argument exceptions were caught. In
MB-19557 we have seen other exceptions (std::logic_error) thrown
during DCP passiveStream handling. So as to not terminate the whole of
memcached, we should expand the catch block to anything derived from
std::exception - this will result in the replica stream still being
disconncted instead of killing the whole process (and all buckets).

(Note: Ideally we'd also print the backtrace of the where the
exception is thrown, but there's no simple, cross-platform way to do
this so for now the what() from the exception will have to do).

Change-Id: Ia36ac9a4775354e4396323af533dbccad83430d4
Reviewed-on: http://review.couchbase.org/64076
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>