MB-19503: Fully restore server cookie API in unit test
In the regression / unit test for this MB we interpose our own
function in place of the normal server cookie API function
notify_io_complete. However we do not correctly restore the original
function when finished, which can cause other subsequent tests in the
same binary to fail.

As the cookie API is accessed via pointer from server API we need to
take a copy of the cookie API struct, and restore it when the test is
complete.

Change-Id: I71045f3d7bd4d181db43876954b99c3ed0db688b
Reviewed-on: http://review.couchbase.org/64122
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>