MB-19718: Remap exit codes on win32
We're using the WIFEXITED/WIFCORE macros etc on unix to
pick up the exit code if the process didn't exit cleanly.
On Windows on the other hand we're just using the exit code
directly which cause it to print out "UNKNOWN" instead of
"FAILED" etc.

This patch maps EXIT_FAILURE and 3 to FAIL (according to
https://msdn.microsoft.com/en-us/library/k089yyh0.aspx win32
will call _exit(3) in some cases for abort())

Change-Id: Iaf0b381aea9946ca8b17a3905525cd325248cffd
Reviewed-on: http://review.couchbase.org/64281
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>