MB-19695: doTapVbTakeoverStats: assume zero deleted items if exception caught
We have observed instances where doTapVbTakeoverStats() is called when
no couchstore file exists on disk for the given vBucket. This results
in an excpetion being thrown, which is caught in the worker runloop
and disconnects the client - which is ns_server, resulting in
rebalance using TAP failing (e.g. a 2.5.x upgrade).

There has been a number of instances of essentially the same bug
(MB-16657, MB-18679, MB-19567, MB-19695), along with a number of
patches to fix it, however we are still seeing the same symptoms.

Given there is clearly more to be understood about the various
scenarious where a vBucket file does not exist on disk, this patch
reverts to pre-Watson (specifically http://review.couchbase.org/56237
) behaviour - if CouchKVStore::getNumPersistedDeletes does not find a
couchstore file on disk then zero will be used for `del_items` in
doTapVbTakeoverStats.

Change-Id: Ida47ec749546fc15aa66a9dc6d75ee31c767874e
Reviewed-on: http://review.couchbase.org/64297
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>