MB-19756: Check if bucket deletion is pending in conn_nread
When deleting a bucket, we need to ensure that all connections into
that bucket are closed, so we can safely free all bucket resources
(connections may have references to Items from the bucket). To do
this, we notify all worker threads, and have each worker thread signal
all of it's connections (McbpConnection::signalIfIdle). That
essentially causes a "wakeup" for each connection, and each connection
should when check if the bucket is being deleted (is_bucket_dying)
before continuing it's normal work - if the bucket is being deleted it
should immediately shut it's connection down.

However if a connection is in the `conn_nread` state (waiting to read
a variable amount of data from the network) then we do consider this
to be an idle state, and hence do *not* signal it, nor check if the
bucket is dying. The result is that bucket deletion can hang
indefinately if one or more connections are in the conn_nread state.

Fix is to include conn_nread in the set of states we consider `idle`,
and to check if the bucket is dying at the head of the function.

Change-Id: I0945dbbcb45ef92aa384240c354e676661c7bd11
Reviewed-on: http://review.couchbase.org/64539
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>