MB-19757: Update the settings read from the config file
We cache some of the tunables in std::atomic variables
hosted in other variables than the "settings" structure. We
initialize the variables during startup, and may be overridden
in the configuration file. Normally the reconfiguration callback
is called if you try to reload the configuration, but the
initial load don't do this.

Change-Id: Ifaab27a10d6879458823a00db5d710c383c3e46f
Reviewed-on: http://review.couchbase.org/64643
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>