MB-19757: Update the settings read from the config file
We cache some of the tunables in std::atomic variables
hosted in other variables than the "settings" structure. We
initialize the variables during startup, and may be overridden
in the configuration file. Normally the reconfiguration callback
is called if you try to reload the configuration, but the
initial load don't do this.

Change-Id: I721f0847d48e9aacff71dfed76179c208319e3ae
Reviewed-on: http://review.couchbase.org/64644
Tested-by: buildbot <build@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>