MB-19815: doDcpVbTakeoverStats, addTakeoverStats: 0 deleted items on exception
In doTapVbTakeoverStats() we set on_disk_deletes to 0 if no couchstore
file exists on disk for that vBucket. We need to handle the exception
in the same way if it occurs during doDcpVbTakeoverStats(),
addTakeoverStats() or BackfillDiskLoad::run(() calls. (Similar to
http://review.couchbase.org/#/c/64297/)

Note: there is more to be understood about the various scenarios where
a vBucket file does not exist on disk when stats call is made.

Change-Id: Idde212db8ed5d7ed9a0eca02805a7ccc5a34e0b0
Reviewed-on: http://review.couchbase.org/64666
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>