MB-19748: Add naive DCP producer to EWB_Engine
To allow memcached tests to be written which involve DCP
(e.g. checking that the DCP state machine in memcached is correct), we
need to have some way of generating DCP mutations. This patch adds
very basic, fake DCP support to ewouldblock_engine. Only three methods
are supported:

* dcp_open - Always returns success.
* dcp_stream_req - Always returns success.
* dcp_step - Always produces a single, fixed DCP_MUTATION message, and
             returns ENGINE_WANT_MORE.

While very basic, this is sufficient to allow a client to request a
DCP stream from memcached, and for us to pump infinite DCP mutation
messages to it.

Change-Id: I371f1e245664e940ccd94d8ca82debb58a1fb165
Reviewed-on: http://review.couchbase.org/64716
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>