MB-19748: Forcefully wakeup connections in conn_ship_log with full sendQ
During bucket deletion we attempt to signal any idle clients, so they
can be woken up and the connection closed (releasing any resources
etc). To do this we change the libevent mask to enable both read &
write events, the rationale being that they will either wake up if
there's data to be read, or it's possible to write data.

In the case of DCP connections which typically just send requests at
the start and then nothing ever again, we are relying on a write event
to fire to wakeup the connection. However this will only occur if the
connections' sendQ is not full - if it is full then no write event
will fire and the connection will never be closed, resulting in bucket
deletion hanging indefinately (or at least until the client reads some
data and the sendQ becomes non-full).

To fix this we simply "fake" a write event after changing the
mask. This will guarantee that the connection receives an event and is
consequently closed. Note this is safe even if there's no space in the
write buffer:

- We expect to never attempt to write again (we should check
  is_bucket_dying immediate close without writing).

- Even if we somehow did attempt to write, the write will simply fail
  due to no space, and we already handle this case.

Change-Id: Ia71838d8e293e66dbc8f48a2ab07adaddb9f5209
Reviewed-on: http://review.couchbase.org/64717
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>