MB-9897: Handle slow stream by dropping cursor and switching to backfill state
There can be slow DCP streams which can hold cursors on the checkpoints
causing the memory usage to shoot up. This can also result in deadlocks.

Initially cursor dropping was implemented by closing and re-opening the
slow streams. The re-opening of slow streams caused problems because
ns-server also tried to re-open the closed streams.

This approach tries to solve the problem by switching to
backfilling state from in-memory state when we see the memory usage
in checkpoints going high due to slow streams.

This switch from in-memory to backfill state does not interfere with
a snapshot that is being sent to the client. The change in state
happens only after all the items in a current snapshot is sent.

Hence clients work with the existing DCP protocol without any change.

Change-Id: If4c128df60fc0249cadf08158a04911120de4c99
Reviewed-on: http://review.couchbase.org/64375
Well-Formed: buildbot <build@couchbase.com>
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>