MB-19821: Remove mutex from is_bucket_dying() common path
The is_bucket_dying() function is called from a number of places in
the memcached connection statemachine, to check if a bucket is being
deleted (or memcached shutting down), and hence if the connection
should terminate immediately.

A recent change (MB-19756,
http://review.couchbase.org/#/c/64539/) added an additional place
where we call is_bucket_dying - in the conn_nread state. This has had
a significant performance impact on front-end operations; as
is_bucket_dying needs to acquire a mutex to check the buckets' state;
and this will be contended if multiple front-end threads are
operating.

To address this, modify the Bucket class so an std::atomic variable
can be used intead of having to acquire the mutex. Given the state
field will rarely change this should remove any contention.

Change-Id: I356c2bafd656644c51cfcaa12c6967b9be182345
Reviewed-on: http://review.couchbase.org/64789
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Will Gardner <will.gardner@couchbase.com>
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>