Perform an explicit commit after call to flushVBucket
Although a flushVBucket writes the item to the underlying store,
the item is not marked clean until an explicit commit is called.
If the underlying store is couchstore, a commit is called with
a flushVBucket but in the case of forestdb, a commit is not
always called, hence call an explicit commit.

Change-Id: I577877d1665137717c0d7db1979d1bf1da5486e6
Reviewed-on: http://review.couchbase.org/64861
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>